pyproject.toml
[tool.poetry]
name = "fake_data_generator"
packages = [
  { include = "fake_data_generator" },
]
version = "1.0.0"
description = ""
authors = ["Alexander Maksimovich <cahr2001@mail.ru>"]


[tool.poetry.dependencies]
python = ">=3.9,<3.13"
pandas = "^2.0.2"
numpy = "^1.24.3"
scipy = "^1.10.1"
rstr = "^3.2.1"
loguru = "^0.7.0"
psycopg2 = "^2.9.6"
faker = "^18.10.1"
lxml = "^4.9.2"
networkx = "^3.1"
matplotlib = "^3.7.1"
clickhouse-driver = "^0.2.6"
clickhouse-sqlalchemy = "^0.2.4"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
faker = "^18.10.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
